Create newversion.c
the reviewed version we worked on last friday. Sadly still outputs a wrong dS_avg. But does react to increasing and decreasing. But when the heartbeat stays the same de dS_avg stays the same value, a few points below it. I have not been able to figure out why.